fintech_ai/backend/requirements.txt
# Flask web framework
Flask==2.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Werkzeug==2.2.3

# Database
SQLAlchemy==2.0.4
psycopg2-binary==2.9.5
alembic==1.9.4

# API clients
plaid-python==11.3.0
stripe==5.0.0

# AI/ML libraries
numpy==1.24.2
pandas==1.5.3
scikit-learn==1.2.2
tensorflow==2.12.0
keras==2.12.0
matplotlib==3.7.1
seaborn==0.12.2
openai==1.12.0

# Utilities
python-dotenv==1.0.0
requests==2.28.2
pytest==7.2.2
gunicorn==20.1.0
PyJWT==2.6.0
bcrypt==4.0.1
